requirements.txt
fastapi
python-slugify
uvicorn
requests
pydantic
sentry-sdk[fastapi]
logtail-python
google-generativeai